// app/(saas)/usecases/new/types.ts
export interface ModelProviderOption {
  id: number
  name: string
  tooltip_title?: string
  tooltip_short_content?: string
  tooltip_full_content?: string
  tooltip_icon?: string
  tooltip_rank?: number
  tooltip_rank_text?: string
}